cleaned code, moved and completed test
