+++ conans/test/functional/generators/cmake_find_package_test.py
@@ -1,11 +1,15 @@
+import os
 import platform
+import textwrap
+
 import six
 import unittest
 
 from nose.plugins.attrib import attr
 
+from conans.model.ref import ConanFileReference, PackageReference
 from conans.test.utils.cpp_test_files import cpp_hello_conan_files
-from conans.test.utils.tools import TestClient
+from conans.test.utils.tools import TestClient, NO_SETTINGS_PACKAGE_ID
 
 
 @attr('slow')
@@ -250,3 +254,69 @@ message("Target libs: ${tmp}")
         six.assertRegex(self, str(client.out), "Libraries to Link: .*Foundation\\.framework")
         six.assertRegex(self, str(client.out), "Target libs: .*Foundation\\.framework")
         self.assertIn("Version: 0.1", client.out)
+
+    def build_modules_test(self):
+        conanfile = textwrap.dedent("""
+            import os
+            from conans import ConanFile, CMake
+
+            class Conan(ConanFile):
+                name = "test"
+                version = "1.0"
+                exports_sources = ["my-module.cmake", "FindFindModule.cmake"]
+
+                def package(self):
+                    self.copy("*.cmake", dst="share/cmake")
+
+                def package_info(self):
+                    # Only first module is defined
+                    # (the other one should be found by CMAKE_MODULE_PATH in builddirs)
+                    builddir = os.path.join("share", "cmake")
+                    module = os.path.join(builddir, "my-module.cmake")
+                    self.cpp_info.build_modules.append(module)
+                    self.cpp_info.builddirs = [builddir]
+        """)
+        # This is a module that has other find_package() calls
+        my_module = textwrap.dedent("""
+            find_package(FindModule REQUIRED)
+            """)
+        # This is a module that defines some functionality
+        find_module = textwrap.dedent("""
+            function(conan_message MESSAGE_OUTPUT)
+                message(${ARGV${0}})
+            endfunction()
+            """)
+        client = TestClient()
+        client.save({"conanfile.py": conanfile, "my-module.cmake": my_module,
+                     "FindFindModule.cmake": find_module})
+        client.run("create .")
+        ref = ConanFileReference("test", "1.0", None, None)
+        pref = PackageReference(ref, NO_SETTINGS_PACKAGE_ID, None)
+        package_path = client.cache.package_layout(ref).package(pref)
+        modules_path = os.path.join(package_path, "share", "cmake")
+        self.assertListEqual(os.listdir(modules_path), ["FindFindModule.cmake", "my-module.cmake"])
+        consumer = textwrap.dedent("""
+            from conans import ConanFile, CMake
+
+            class Conan(ConanFile):
+                name = "consumer"
+                version = "1.0"
+                settings = "os", "compiler", "build_type", "arch"
+                exports_sources = ["CMakeLists.txt"]
+                generators = "cmake_find_package"
+                requires = "test/1.0"
+
+                def build(self):
+                    cmake = CMake(self)
+                    cmake.configure()
+                    cmake.build()
+            """)
+        cmakelists = textwrap.dedent("""
+            cmake_minimum_required(VERSION 3.0)
+            project(test)
+            find_package(test)
+            conan_message("Printing using a external module!")
+            """)
+        client.save({"conanfile.py": consumer, "CMakeLists.txt": cmakelists})
+        client.run("create .")
+        self.assertIn("Printing using a external module!", client.out)
