+++ conans/test/integration/package_info_test.py
@@ -88,57 +88,3 @@ class HelloConan(ConanFile):
         client.run("create conanfile_consumer.py consumer/1.0@us/ch")
         self.assertIn("intermediate name: intermediate", client.out)
         self.assertIn("dep name: MyCustomGreatName", client.out)
-
-    def build_modules_test(self):
-        conanfile = textwrap.dedent("""
-            from conans import ConanFile, CMake
-
-            class Conan(ConanFile):
-                name = "test"
-                version = "1.0"
-                exports_sources = ["my-module.cmake"]
-
-                def package(self):
-                    self.copy("*")
-    
-                def package_info(self):
-                    self.cpp_info.build_modules.append("my-module.cmake")
-        """)
-        build_module = textwrap.dedent("""
-            function(conan_message MESSAGE_OUTPUT)
-                message(${ARGV${0}})
-            endfunction()
-            """)
-        client = TestClient()
-        client.save({"conanfile.py": conanfile, "my-module.cmake": build_module})
-        client.run("create .")
-        consumer = textwrap.dedent("""
-            from conans import ConanFile, CMake
-
-            class Conan(ConanFile):
-                name = "consumer"
-                version = "1.0"
-                settings = "os", "compiler", "build_type", "arch"
-                exports_sources = ["CMakeLists.txt"]
-                generators = "cmake_find_package"
-                requires = "test/1.0"
-
-                def build(self):
-                    cmake = CMake(self)
-                    cmake.configure()
-                    cmake.build()
-            """)
-        cmakelists = textwrap.dedent("""
-            cmake_minimum_required(VERSION 3.0)
-            project(test)
-
-            message(${CMAKE_MODULE_PATH})
-
-            #include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
-
-            find_package(test)
-            conan_message("Printing from external module!")
-            """)
-        client.save({"conanfile.py": consumer, "CMakeLists.txt": cmakelists})
-        client.run("create .")
-        self.assertIn("Printing from external module!", client.out)
